Rename feedforwardNN to FNN and update cls diagram
