--- doc/rtd/content/04_appendices/appendix2/sub/core/mlpro_sl/images/MLPro-SL-FFNN_class_diagram.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2023-02-28T09:12:09.900Z" agent="5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.7.4 Chrome/106.0.5249.199 Electron/21.3.3 Safari/537.36" etag="5BVcHi5dS7jLrXYfdMzr" version="20.7.4" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7V3/d5o8F/9rPOt2jh6+ifhja2efPrXOt7bbs/3CiRqVFQkDrHV//ZvwTSARUUGt4nY2iUAgn3tvPrm5uamIrdn7nQXM6SMaQb0icKP3inhbEQSB4yX8HylZeiW8IMpeycTSRn7ZqqCv/YV+IeeXzrURtGMnOgjpjmbGC4fIMODQiZUBy0KL+GljpMdrNcEEUgX9IdDp0h/ayJn6LyYGr0F++Adqk2lQtVz3X3kGgrP9V7GnYIQWkSLxa0VsWQg53rfZewvqpPmChvlxv/yhd17lu3//Z/8BLzcPz93vVe9m7W0uCd/Bgoaz860nnGxqtz/q7T/yozh4Qsor/6vKK96934A+91vs3nCgZSIdGI7tv7izDJrTQnNjBMkduYp4s5hqDuybYEh+XWAJwmVTZ6bjIx5/HWu63kI6stxrxXGd/CHlyHAi5d4Hl9uOhV5h5BfZ/eBfgK5NDFymwzFugJs3aDkaxvjaL3aQ6d+27z+q32TkRPiekIQNrciH0GKtgGgGHWuJrwvu0qz74uBrhBiIx2IlXo2mXzaNSJakiL5Y+yI9CW++Qg1/8YFjg9iymrdfe+BZfFhU1dYDqIPxTx/3KIZtiDESuDayFsAi37pwbgHd/eIskPX64ZHNBiNDArIjqwiNAyK7fJovmz+eUU/4zT1N7349qFqvSq5PQNufD6rPyNSGFfEa/3ILHECAhERZOWAQsEdhGWkoC71hE2yRn69sCL3LBuMaxgSCmf15P0kIjSKfCrP3S2CB+d2AXy93CD+f5hDcJC5FDCcWGGkw9lvb/eDfihIrKS5UvMQwFwJDqMIL8xcqfpNQ3V13S6koVCp4iU/IhaDQgiHVWYIh5iAYXGfwwPH8L9hptf88v1bHxrTB6Ef6nesRMB3tDbbnxtDRkFEhsoq5k8iTJr3CdmSm15In0TbFXmgzzChgAj0xAvkQEuKxFvThVNNHHbBEc9IOtgOGr8HRzRRZ2l98WxDKnwMsxyejLvarM/rkSl+KLWjjc3oBunxY1AG2458zRLoOTFsbhOI2A9ZEM26Q46BZcKOYcjCkPyrjCXEdAaiMh0w9GSpwMGZL6Nj9kKv8lm0nlCD4vSgJbjbiAiwwDJsgM+S33kyRX7+6JzwUAMYEN3lYXyj2KYaUZ9UnyPHqgI7FzAAOvCGg2ZTWhK+6u4WVKEVqqc8/e1+J8pBD7lOgMYSkrTTrqt/5/InSHQyIk9U02sTkGZOOe86ttCp58huEFCF87Vh3JXSqjUbQcCXYwazBE3IiwybSCP/HD1C/wX9xu7a4Wr1Sxw/Uwsf86hj/JadbWDoNLMJAc+UMYg1aQKJFDAlMtT6bJXAZ71k3CVxSAHLj3TzdkbbU7vVjBOeK2A7/nh+yiR6Vw581vWY+iMvy4RC/B2Pt+1vrYdB9cB4t+Pj28vRSbVCAq84UdxlT4pY4N3SLQrEpHBnFJo3iDJgmbm1bdRBhESWU2aDkA//R8WwwPUJegTlBqFTLzFhK0pH1kqcHIKoBHXXmuaVLHLPh2DggL2I+Mk2LKPB0zR0LRl0G2/oCZhgjcrsAymeC9G2Vp+AWabhFBrQ6GEC9h2zNJePireWdm4D8aKg2M3IfpSDlZOimqhmao6pX5IXxd3PuqLbnErp+9HxDWClWzoJKS6hcc4Qb7/HFVPGQ/7A1QR3O1KFOGvQr/kqALaK2FZEk8lhEDYP5eAwt1fb8IsXUAUYAgzJAiOj1szX3kOHyr+hNs+fYQvwlQ2i/ujbQ7WIkQUeTCTYfuI4OmtRaaufbnXrd6RRQ15cvpvq6ANaE+OtpP1rZ6bHNo8BlJC+8mIOBZDNRkbaQNnTmpsdfrs4PzHXD/7gjcwTHYK6T80fAnrpOUr5ASRCljMPLPNzozJ6S9v4B4u2LdJPeLEtenUnQUW2+6x7/miqZVSRTiu7hZ38akRjecxPrwiSzkVEy60VxOJm2UMWK5mFllC2vwlkLa9QGr7G068xxUXIuBZZ1k5zLOVhgdlfM8PC5gq4ig4w8i5f3/MSYpTdnb4LXMYsUuTaQ61EoTKibGYU6j7ANplAHQSIMoR6PV1J9eWbvNASkntXqFcY76fHHBDoqGA7nFhguz3D8URiUckYoBWl/KL+rty+L6ujvnVj9A779KyvvN/ehJzwCFhxNYBA2g9tkiibIAPrXVWkiCGV1Tge5nlGiEb+h4yz9uBgwd1DcBQuN0TUJucaHAx2RKBlSRAJL1o7Z0mJJbDS3hv7Tv/SE/mw0M2ba7fOi8bu+NHgUAOMACwtqSmv455EWSIXPgjpwgypWJ+2HTdpTx6Kknol0EgeRwPkhUUHJSYVCCXLlpEOhxsoQDpmhUAOlToLhNoc7pYVKbSe+u4dC8eHqh00GpJHSF3y4UCimttDTs8kQmX7n/AJjaElLNSXZ3ZwHDIBhPjHNMSjsLmhqLx9Q6xndQjl4hXo/ndkIvshq9/fvryPUNdHjI0NFKUwn2Kqbe5rPcI2XD0wluoqKGSOtJOwqw8xJrPUY4Yqt3JuKJf6yHgxI3FcMuh35zxx5UuOOTjguWsTZwUK5oIyP/ixP/P/dWw9iWARnkQqr3m3IKI+XzHf6Fh04wQzOPcO/l5W8O26HAaPMe5/Eg1AXZynZ1DK4a461zKY3lVkvmvp+rHfBVsNg1uY/J6nJmgyuXJeOO4UZ+fZ5/QP4b7vxLZhw4f/BjPRpxsA2veagi9a9Vuy0e/JypoUmmMTZmW6ba5FbPdYSbQS80OaDPsATBCNiR/D4RuDmZEo4bDMP9u0ENK6665DaUv6204o89DBWEnt44DiWNpg70G3LNv4XW8spGhGXQX4odeGCyOR8oGtDYptXlba96gLEBuQ5tJnpOTvJQsYtrVKegO4IH6uK1GaOGKWwJnyvZvxREifkg40LTFszptDS3ObeBNIUkHfX7CRMuBO24HpY8tWE2CuoVNvmiibrEdTtdMSXfsxMMQNyGytoz3b02U9HAegm5XZ55bhcRV7/FERrTTvkZWSzPix78BsZQGUaB7OWZMadMZTTXJY516eS9Lb4bZNw3ySdNevdKqnDjM3joii/V5q1ZvQjHZLu05Ect9CErqeEI6wm6aKl/acRENcNQgM0O4nfw0Fw1sFy3Ffr+ha9UQYvbI3SRl+rj2CVlwWxxuNxWPBRYkO2Kh7F0aO2Ol/jJF5efRio1lNA9R+nR4bvkfEiVxM4hcOdpuL+aSacclyN40SOrysN/K/MJfzDnrfav+VKbqha6nUskJxSJ8u28R++IW1Vjefs3lxNMzHaR+OxO6u3ugSLGlhGzvCdGWubR+RqcgSnZhBQvHrwhL54FezqDWRKV51SqHuveyAaNbbQ7HR1iit1itYpUaiJiiRzIlYD/G9iOiwvlUqvZVuNWkYwSNOvvGSeDrqihLycEtpnSijf1fF7cpeY3oirj7DRk8nVoueLIq2+Scn/0LNFzHam19dSunJB0wup5mSfOaMMslbUZAOdb6y0hltYw/FYYFvDkTyQ6/KpWkNeqAnNKDkprWEGXaGjWEtruMacfFBryFjxXJrDE06dlI85xEanJjYiQ/FGaQ6zKEu5rnyjLIYG5fQNIjPUU6IQ/dhxr2nxrNG417TA5cPHvTr9X/ff/vtxZ8AOqr48WR1J+CeIfTkxaOC75vxH7l+r+0c//drI99v36MEyODBwe0QuIoc/wwrxweoy9yi4rngxYDZ87mKwk/uY57l4V9GUYw5j2qunpJ6/t4M5rU1jwdl9TI+ApZGZVT84Oygpg7MvMDibYyTsPv/gbPYyE0pbknkq+51qoCyXEKWdalNOMkqb+cRllHb+oB4wSjuNBUQwxX2Zv9jTrpE08H131eeVX1zzsv2UXdz5d3FSYuOCy1x+xNQauoer1WoX0JGlmpCT7MiYTyxQ8F1yR5YPqAfsyJjJ0U/Tj7CdbY6O7ZkvebSx/V7QZOIYPX9DmZJYnD2xSG7xcJG8gqkpl8orUs3GSfIK9hZIJa/IHdRj8wrl5HjFfnMGzJes580X1jn75bgtVpJZi9ZE3W4/q5CoqCnlO0uQ1orRDRPb191uxd2B56Q3syqJTiFER+AS3ENRsuYTllP6qPPgOnQAe3KWgNpuFCvTJUwXpBqXj8OGyqjr/EE9IBtibuws0mP5C81Gvga5rKCvRThrksA88j0yIRbY+1idObTJ/MX5A5s102seOwmwgaUXSYy9rjWeuDjMWhykeg0Lzhj/Q6g2L2Tc1iwPEWD7l+gu+AxHs74Bi3rJDzPq3Qsb2uw+dnpk7OgOIsvh4gUMF5O7d2ceLl6AZ5wmnsmMn1hdLnZ02Mwsa6cyOiyXXRSA6rGd5bx8Cfwi2JLr8ARjJ2e4ICYCuXiu0YxCvfmKptxIyEYB7nPGTmdPN+2uF35YkqBLJUE8JzZLFuSrCJ06LMmCsMpUSJI852K5UGhHPhAZoqcOSzK0N6xHZ0P0DMjddbfsw86+D+Ol5OowMevu6RfQh9Fu6AsOcgttxAfqrMoEMgXAeuzOSmBQSy92qRx6XUi3RQ+9+GbpgA70g6boyaFX64LDk0Lz8XH6MYGOQCv7sb1hPXo/1qBAPEMXdGCNPqwLWvDTPx3NBc3e4ZuyCP1OFzoLZL2eEv8RuRPnP5vzie7Af9rup8hhu5hYNcDKmSdyrD4qzZx9OP7D3iecQX+SSV18VbkEEpRqPjJ3lozNvdnylQMHYj4xPeWWJLUVsR3+PT9g00MA8we8TpPewwJect78QVVoyssENQfKy3xgeppBVTVDc1Q1Etir+ptEeq3SRRhjsusk2Q2OdwlqS6hcc0X8a6qY0mR/AoHb54upTpcmtDDIADcuIV/X/5CCHil4npuuSHl151vxly+m+rrAPMo+v0D5ohSHERLPVJw079JemkNPRVz0apb8EeYZy1mYEOexnIX5yPRkRbno4aASkFXJ81j0wH5k2nP7PHX3ggw3iHSPDDgcQtsGpEqyH7M2diXEM+q4lT+ZqgEdzzR8ctvevw4RsRnBMeZNI3oiczWQJvSHtZ1ihGntMJCO7aUYcUQwaFpcHoUCh9KNRFbfukwTX56166IsFCUFRwlNHQF7GmIfJJUOvrsppcME05uSSlciKaUjGaZzTyqdqkNRfyP7RH9EWXQqCbmR2N1AkmI7etbr8Ttm3TROTkyBSc1sOy3m5oGkyfw3wx2RCZwOgWVg/fU6qKDvwo9mjNxjj2aTAguGdolw3Rq9a+w3ZwqthWZ7HJxD43GyktKOMeyYzOAzB7ZjR0lIs7cdy9seNQ5jZpqcGDcrcWngGrtZmWT2cLmZoL8FW5mgIU5okmz/DrImCc1oJ8nXONxJpHeT7lEPWprvLNhTVu/BWPv+1noYdB+cRws+vr08vbCSNqaNVIpPwkQH2HO7iTF1K0XKJsdb53OSlPgWsXU+Nn2492wge8H8UbaNYWclyDjBnBCe3aSY2RhcQcJ4qIxgjcS+IZJUT5ft5AWJFGKHzzm2HpYYXbwnc84m0oH7cGXmsQ8Qyrch1cR6vV0fypfMPCZdZJJVpsps3p4kokLnN+WZSeS26QKOEM/H7q4pXCnsznxqM39Qj51v7ChOuxNlYFlXnTf3ZGr7JYijlPCx03tG1nBKyAj+XrKPC2AfSiJ48zIT2TAVZPMK7t4y0JczzmizS1clnjb/KNdw5w/qsfkHPRvTQ4hcNdSBTSZ/qxVBBjOiaxOvHbi+A3Enm0T+A8ye5N8PiMnNXcN+IWqYJQaeST9iboDyR/F8nyijVDIySn6Nkh6GUtJBWl5aoDsLX4ZJpXdU8soL5JUXmhuIqSZ0oFuSWBKFqQ6A7WbVP/s8QbuQEeW0GSYjCWZJMfdF9dgUM8jJvA3HvIUOxq4kmUySyXOMuY4Ds8zSbxlpjKyOS2++/Vg0k5F+stXthq7LVpkD5TIp5mXmQGGryDbOyzPOhrILCVk3hj4Zblm6LwuA9ejksnRg5sstWQvaDssthaPsdH6q3JKx10sh4Yv7aSHtm4mEGAV+zEhRyTQvkGmWIXorG0c7ulKdmWXAXordOV3OGbxIyTnzhPXYnDMAsXRo5kQ6ReXonJMeRvRNONRwV6BB212PmlxgijHmjPlsgBkJWV5KjuDcwrqAH0AHS7f4amp+Zl8JhpiVAVe/BG7sL1zIdCUyHW2GGcjqrI8kU+xeqTAKklztrDASkoUJHPbMyIIPLYScKFHAzT19JOkUcOH/AQ==</diagram></mxfile>
