--- doc/rtd/content/04_appendices/appendix2/sub/core/mlpro_sl/images/MLPro-SL-PyTorch_class_diagram .drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2023-02-23T10:36:59.465Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.8.10 Chrome/106.0.5249.199 Electron/21.3.5 Safari/537.36" etag="X9DDp72_ixhtLj5YMGHW" version="20.8.10" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7V1td6K6Fv41Xadz1qqLd/VjtXVO79hOz7Rz5vZ+YUUIwhQJB3Cs8+tvAokCiYqK2hecWVZCIGHvZz/Z2ezAmdqfvHyOQOjeIhv6Z4pkv5ypV2eKImtyB/8hJfOspKtJWcE48mxaaVnw4P2GtJBVm3o2jAsVE4T8xAuLhRYKAmglhTIQRWhWrOYgv9hqCMaQK3iwgM+X/vDsxM1KFVU1ljv+gt7YZU0bupbtmQBWm15K7AIbzXJF6vWZ2o8QSrJfk5c+9In0mGB+3Mx/+MNn4/N//o7/Bd97Xx7v/rnITjbY5pDFNUQwSHY+9f1TMrHhd8O8+/nz2kZ3Ibq9vVCyU/8C/pQKjF5rMmcSHEdoGlbsAe3pLxgl8EWkXzBip11KEGMPoglMojmux47SO1TqFHeG1M62Z0stal1ax80pcKEuQJEzXpx8KRz8g8pnC1mpAlkZPm625yAsFXKNPorSPca/U4KK3pmiSuknXyTFzEpYmZzfbYzp3/TUo4IyWC3S4EV2mktcQdbCF/4UQziGgZ3WoOeKymfHchgJyrLrKXWEO7hKySbJOE5RMpuu1BBd6NrrE11LHIJA2BrtJ2kpGo/Osdrwf4wWKffr0+oO0KvdeBVCdeG/YBLiH8EoDjNx8EWrLqtQ7YZcXBihcQTjuNJpay1Km8dW4tkg8VBw5A58g8Cep6MM/prGMCezTO3bAbRouqs0tSX+trOKOuywUFLoPEiSyBtNE5jKckAGPpi4yD7/VKeW7uCMYHI68j2LcPOy0UHWHNPYiPTDm4Q+nOChBtpbs1KdCt1RfaIm1oo5R0qLlvC5usWulCrUo5tUMQMvcGHkpeLepCQXkGv34rKa8CAcwdVqqdcSCpdgcrKtVZuiLpjb2QhFf4QS7AKlwmLyHOT7/noMgBeptMslF3GVu/zXAK0VcqiLZKt2tuRtY7+ZlLvJBPumVzJhDt8bB/i3Dx2yh/jWHp7fXNLiBBGJz1ws5YcQWOQ0MzyVw2XYbw9sSDxaiYqjv5AFloBhSKQ8TiL0DAt7UtngPWy+Q3rheL6fq+Skn8UF5CcGa+cZ5YkBPwHI+/edbqub/2jHdPc1zt2/giF2qAk8iVdTUhzec0nmq3hr5CPruahEH4ygf49iL/WI1Ksou4KFNoel/RPPtsmZN2l/UQ83P/B81hxRNp2Ly8rWWoJ2YU69cpJ2IRuK2pLxPIx9OoU524Uis1lcTqu63JI02Vh+BFrV1yiVduceeal9MzRJLUXqSHjQ7KT/uoWeyIrUkiRVwtPKNv42JLl4/hhNIwvSU+bn0KVWdB0DUuroitIm/+S2tlUzCYjGMNncDJMjOw45TgwLh2CogXmuRkhOEK8Wjyq1jJyeul2j3PGSvWQNLK1nof3dDUrnDOomGx6IRTkRmrxem5Iam+JtSlVaakczJBWbAf7WDmNS61vZ1qLmOR2ss6+6MG9sjq/FM2/igwAyiNE9ag6oFnGEopWDv+V6vj0EczQlqo4TYD2zrZ6LIu83Pi1gVoN3RwzHeJaRr/FAjqRojyAJV9wzdMuLoiGIE1oHu1Y+CGMvi+mRKhOsDi/ooSTB1kxPVPBCCj5F5nqwsKzIy+hY0LJETsqoo2s6OeE4ArYHC67Nwjdhkh0cy3cp2I26/CgF8Gm8/WpSK19fVXnzLSNfYL7fSIA1GPtw2RVM/EXgCxrH45TArzKKzQEfIzAACewRfcaHsJX2ZlvxvdROcoMAj6Gq7I6n05YXjB+JCV1dyMuSYXpgaoC05BuVi5riOQEsjC2tGYXoiIy7rffO9Ku0JMIwDXCPgZcCDmJTmsE4qQ5FozIU50U1boe1zmqo7eVSdxo23IcNHUcRs6FtjAzdeK1sKCstpZt3Tho2rGAr3YYNN0GxUxmKr5IN2S3yhg53okMbwI4jpEPD6sCR80rpEJNOS23npuLthg6rGIu82Vg+Oh8uCOX1E2I/6l5d34NH9cvswux/ATpwni54Fd/PH1FkuQ/DO5jMUPSc3ReTiWTPY781IelKcSu3m9yhT8ghrSBo3SJ76sNPr4lUNel1k2oFH3MTcQpId5B+hEDeKo1otY+5uA3AgK3yVKZKAiqT1c1hsbdDnUKzUjiz6puPT/fXuCzdlP4om9kfnMnQ21KV7kUViVLjiZIUIXys46fAdDHnwkBAnkWS7JEEHKmlE7rEdq6TrCm2vQODrmWgygQqULsQZmW118aZfFJa37y7vM0p90wdLP6/P8VytzUlaQXT1KNwgx8xj6vwChHlD+QH1aPUrlJNqYdyfPi5voOiGYjsc3K5pheExLO4zBybRxjEGO6K9ClLJiyWNua9FxJkqSKhr3MbqmLhIbDu/3VvnC/aLwveG7fOl0n7gg8K0NH55ut1lhKTd4JHTsvGsm/1po4DI1qh8Xg/oMeryRWR231PDq/QhESJ+h93iFzLMrU7ujUMkcIO89lYJh4XvcQ0142R/RxXnvWVs0vpEN+hiaZhwjef7uTJ+B2OyWm+4MGg16kGPU05EPY6nApJNg8bRLGkXDRGAfCvl6WlIWlZZ4hSmiAs8RMmyZyOkmCaoCIfYSFG8/+S41tat80KntKCbkdlBVcvtIlsa57fuoeRhyVARnVxCuhKxa0bo7LEnjXionleWWrOZpPemBkVQR8k3q98pQMomc9Oo35X5lqtdrryQcdp4vlxtvMKf8Ww8cU+oi+mV6WstZz1PpyxJl6xcUTUKwPulThjfKZW0RmjKx09SiVHc8MSLHaTsC9uHDtEICWulnq8DoxAYrmly5ek47UfQ8K8C8E3Dmg95laRzduHckD5+CCwbRNm0Z7M5MgG1nwhDpTGCBvl76d8tjTidBFBQZ4Qdqwx4aZjt29aeOzGP89pPJilwjdK30Pp+smVzidDmCZWu5fASTbMeuQhLZdeY+b1aFywqubIGlc4jcdgEmLneHze6Hdf/SpVw+OH0y8fln4YXtogJLGNwTSw6MMyhAlO5XpNWOH9hxU6cnGWXxnC+vu/xaPwPtGKpKaLEYjThwgwE8I/mRW9v3SYivS54KJXmegk1jjvEPXN3vfB4Pqb2R8+LPReito2+n19eU3iLvN3+DjtffRA4fZqPWJmk9hqeadWcN/WjLMnl1zeZo8wEcXLFGmfH+QWbXLclkhExrR8ItFFPOYArSUudqjc7LGY4kjjvi2MUiqlUc0DtQFsErMdIUQM+zGaZpqR6m/olxdPMUX8Jm4AbW4A/PgwSPDRGPuxY9zGEI1bfXP49bN5ORweoK0//wzN5xl2m+N3GQ04GD9WzveT1UNRpMzfIctIMYxgGCGrmN+SD+1+yPTPbaIE2+NBUSvG+g8WJlD4MAEdUBo8nAAPxsnxwPvFDA8oTnhAiJPEfS9u7grsj4buydHAP0LDXDNaNCsGFisGbBC7afxwOW1k1WzogKl/UD9DVU6NHHZiAXK24JEGOseHjn5y6AjuRe4wBjXYOT52OifHjiAANAHhCie2nxZn2MqVN3lMNQXx5ZPDgZ/dmCSXKYsxZaig8Saa3oRhwC+Aa8BQAxiqprUdDgx85MOEZCN9b0QGBmAlPB8cKZ9ywzfpHva67ZV8lY56WR5qg9Z90do+9SoglV8hQsLWIcZnQG5MrR7STg/VdRDmrYtBN42PN8jdE7l61QRSrQaeveupz+iv5Gdbhn4/Gvuzf0Z3giXlDLiOs0SunS1Y+ljqHwzYyw5WuOYF//1wGDEqYuRwY7HGqfq4ixx1PM0tLnJcLHrcc41jOg+DL15CG6JbT6wP+PeyBbLBGuDVvde6SAaLjQsjWSZofSsj00O3fWdAl4GSgrRDCWrVM9ZltdNed8DebxUQA5dfjPToQnJDl75kh9wRxSrE3ygkfiXgSW2JYwII0btUcpjdIbmw8CKVXHKmIF+lyJni9wzUk16o60X16oJbHYus6ONwUPcUHFRgeEYTco4klpQhpokliRUJTN/AXzXTCzOEjfTCQtE10wv/zhGt+CoHQ5FbUu5TctSrvjBCKy22NcpvAFrxToi6CEeT3yRKT4U2xhaHRltbK8FC0nfDV6e8mLujHhdfypvE11tjQXZT4+C4LA2zhrwrLkvp+Xr3yLhUT4HLXaBY/aFxe0FNEfjz657RefwHnXTBnQn+Hnx9unO/vVz8T5/e2D1BNKD4XjspcYmznC6jxxvzMHWdnWKM4O050IJ3kldDQmUHum3o1RzoOgI9QtW+Tc9kv5GjZFW72fc6Q8mb9+rw2sH9GY7ud/VnDKWucQNvRoi8P3VZHRu7e0sWLeLC/wM=</diagram></mxfile>
