+++ doc/rtd/content/04_appendices/appendix2/sub/core/mlpro_sl/images/MLPro-SL-pool-afct-pytorch_class_diagram.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2023-03-02T10:20:27.289Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.8.10 Chrome/106.0.5249.199 Electron/21.3.5 Safari/537.36" etag="su1BAPrf-utFezO-FeDV" version="20.8.10" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7Vxbd5u6Ev41WSt9iBf34Mdc9+5p2pPTZJ92P3nJIGy1GFGQm3j/+jMCCSOEbZzYTnIC7SIwCF1mvvk0upgj+2L2+EeG0ulnGuL4yDLCxyP78siyTMd24A+XLErJ0PVKwSQjoUi0FNyRf7AQGkI6JyHOlYSM0piRVBUGNElwwBQZyjL6oCaLaKyWmqIJ1gR3AYp16TcSsmkptWzbWz74E5PJVBbtuaLFMyRTi6bkUxTSh5rIvjqyLzJKWXk1e7zAMdeeVMy3j4tv8c1P749//Sf/hf46/3T/5b8nZWbX27xStSHDCXty1rd/s1mI//JGX378uArpl5R+/nxilVn/RvFcKEy0lS2kBicZnacdayBq+htnDD+22ReNZbZLDQL2MJ1hli0gnXzL9YXWBe4847S8f1ha0RmKNNOaAStzIYGcSZX5UjlwIfSzha7sFl15MRR7HlHQCm9jTLPiifdrzlFxfmTZRnHURUYuvUTKzPpjbyL+FlmPFWPIVLzAkzKbM0hgOumjnsUNnuAkLFKIvLJm7qCHcYusbE+jItrLXSSbNBNFqmY2tdRra+ja9rW1JU9R0lqaqCcvKZuMj8Fs8B/QYtSuPqyugGjtxla0mgv+olkKF8k4T0t16KJVzVKSfeSNSzM6yXCed8p2p6KiePASEiJGaHLgCnzFKFwUvQyc5jmu6aw0+3YAVV13laW2xN92XrELP1QkSuURYxkZzxkudHnNOz7MpjQ8/rBLK33BDxyT83FMAs7Ny0Kvy+Kkxca8HmSWxngGXQ0Ot2alXRr0ieZrK2KtmmukVJUEeQ3VqjQS7MY2hWGuSTLFGSnUvclIU8TbTvKmmaATzvBqs+zWE5QmjDTd7tSabVUYbecjAv0ZZRACFcqS+ryu1/31OICuUuMpTVZxVWv+a4DWCj3simS7VrYRbUPczOVTNoPY9NLkzBGTSQLXMY74Ex5bExjfnAkxo1zjD1PQ8l2KAp7NAwzlQAZxexJiHtEaQh0XlS5AA55ncHnOMvoTK08K3cATOd7htYhIHNcSRcVRNaA+MFg7zmgODPQBQD2+94eDYf1wDhnuO1q4f4lTCKg5PHlU0zAcPDnj41W4G8c0+KkaMUZjHN/SnBQRkX2ZlS2orHnTeD4jYchz3mT9Kh0Uf01iWRw3thiLm9bWVsKhMqZeOUg7MT3LHpgwDpOHr4zZTixTjuJqVnXNgeGY3vJosaq7xqiiOreUFP4t0WQMLMM3oNP0i39DpSamZQwMwzZgWHkKZ88w1fxzOs8CLLKsj6EbpbguANLwXcs65f/MU2erYhjKJphtLkbqUb5HoyjHyisANbSopUh5Bvlq9djGwKvZaTj0mhVv+EtZwNJ7Kus/3aFczaE+lt0D96goo7PX61NG71O6T9nWwPYdz7DBDeDs7Mel1peyrUctajZY51+7wry3eX4tfyCzGCVYQkw8sWtADXgglK3s/IMpicMbtKBzbuqcoeCnvDuf0oz8A9ki6TXwOJM4hlFGPcUdf1OgPcN8uuJWotusRDcoZyINhFYxSnNSzunxJDMwB0nOKWPgzSIjJQpRYooy9JDTsm1Rhh/gIGgLUsa+67g8w0mGQoKV0KaKTaRmrw8Vuyh+Yy8PSwGfo/uvYwzq6W1bd98m8lvc9yufYE0mMV5WBYhfBX5L4dBPtcRVnlocigGBCWL4nNsz34evnG72lZgUflLrBHQMdWV3GE4HJJnccxe6PDGXkpvixcIBheSr0Itd4JkhOY1trOmFRI8M1XbPj9zLQpIBTBOoMSIF4DC40gPOWXcoep2huFDNuB3W/NVQe1ZI7fds+Bw2jCKrnQ1Db+y53mtlQ9MaWMN6cNKzYQdfGfZsuAmKfmcovko2lEvkPR0+iQ5DhP2olQ69wMfj6JXSIZDOwD6tDcVPezrs4izmZmd573xYEcrrJ8S7JLj9Nf0YfXJ+B/jW+xx9mp2e6Hx4u7inWTD9+O+rcjWgXBczuWaPx9EgBF0PzudRhDOR4MNrYlDHePMMuoklWxj2ujhaUbvVnqHVAWU15y8509R5yzZaeGu4eQrs7dBkqwu17VF6vyy5lmU6k2SLuVvRtS9a1BeiRiOSEDYaHfP2wnXKueiMcaoc3OMkB1fke5Uqrjy6sI7OjH2c0xGdp0wvvniok7FY3ey0pKkiydGRxEUU3o3igvKmAEqctKBLRdE538dlDFyOJ1CSyzffyfuNENMWUcVS6d6g53eDnmPtCXu+ZkK+kCE7UdDUlE5oguKrpbTRJS3T3NCCJjhL/MCMLUQvieaMqnwESswW3/n7AwfCYyH4uxAMfVsKLh9FEeXdon53izMCGuC9evvq90rDreujyjWNNeoSS1zlqsRml964KJThGDHyu55oD0bWF+ZE3FWGVquDrnJHZOn4c0bivHx4Cacc97HYe4zF3K6UtZaz/j+CsQ6Lf+89GHM7A+6VBGP6IpUajIlN3kRQycHCMAZqH3H2hcIhIEIFcQ3sw1VgjFgwbTTfMA5Xfo4581aK7wPQ3bhbRzY/3VcAqi+DoDAc4XK2p3Q5fgOWV+aB4EFv/OcaX+4K22h9e828yrOs37JEAoE1EG7Rd8ejAPpuuDz+UP6uRu4C6o3+DKO7L250fap/NAKzE4ZnZTdL+O9Tz0jv5ruxeMuGwgNb3NIsnqNZCsHx5Li373Pta3WdHt+fffVpaTHD8CeOU5xdz5OAjzZyzdYvOGNg9zMG+5gx8Bs/7Ha7ovMdrN6Y+lS/5hHvfcag4pLO/Ge/8JSBqc8DFUs2ozTDaUYDdRWnPoApglp1WafvC58Jho4D2r31hZbeF0Jn1MPhZeDgvTgcdMqXcKA50/Ggml+CIiZ5P/J9NhiGLw4GfYP8aE1X0Y6F90YQRbQaonxaxNHLgEgmC3GE5vFeIwzTemnkyIxbkLMFjfTQOTx03BeHTst02xO6oB47h8eO/+LY0SfuRqLho4eiAqMxQXmxQspBNKPhnA/OD7Mo+FbOQlNcSb1q1HOFnl4x6nkCJNWvfuyCRYFGu7Ho3ta4bX1Tyf0U80/Xie9E8Q9kQMPhTFM+x4ZizezLmWTeFbV9Dqg2Y7j9D0vVbwHVZuJbphtVVLV/KmM3c8leY27XaxnOV6tbh+kQh5plDrBhVglD8CNh32vX5b5ZV9wtN83yG2XP7Pf6Te2tNXtttzPsxu2z0hE27p+V463dbaAVrzY+G+KY6tdIhpY5MGpHgxG6fvOk+nBtlW0Djys+a7KrdQbHfJMofSm0SbbYN9p8swELw30avvwGvnzfPiy+9GWrt4Cvt8aCcuS+d1w29ngPzSficmg0cDncFS7hdvkh9jL58nv29tX/AA==</diagram></mxfile>
