+++ src/mlpro/sl/FNN/__init__.py
@@ -1,0 +1,1 @@
+from mlpro.sl.FNN.basics import *
