--- src/mlpro/sl/feedforwardNN/__init__.py
@@ -1,1 +1,0 @@
-from mlpro.sl.feedforwardNN.basics import *
